% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{add_area_column}
\alias{add_area_column}
\title{Add area column to object data}
\usage{
add_area_column(obj, areas)
}
\description{
Add area column to object data
}
\details{
Adds new column to the fixations and gaze data frame from the area
}
