% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-analysis.R
\name{analyse_gaze_areas}
\alias{analyse_gaze_areas}
\title{Analyses number of gaze points in each area}
\usage{
analyse_gaze_areas(obj, areas = list())
}
\arguments{
\item{obj}{\code{\link{EyerObject}}}

\item{areas}{list of \code{\link{AreaObject}}. Only necessary if the areas haven't been added previously
with the \code{\link{add_area_column}}. If areas are present, the column is recalculated.}
}
\value{
data.frame with calculated information
}
\description{
the totals int he data frame do not correspond to the total row, as the NA data are omited.
}
\details{
returns summative data froma with columns: area, n, ratio
}
