% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-object.R
\name{AreaObject}
\alias{AreaObject}
\title{Defines AreaObject}
\usage{
AreaObject(name, x = numeric(2), y = numeric(2))
}
\arguments{
\item{name}{name of the area}

\item{x}{x limits of the area as numeric(2)}

\item{y}{y limits of the area as numeric(2)}
}
\description{
Defines AreaObject
}
\details{
AreaObject is a list of class area with given fields. It is used by
plotting functions and mainly add_areas.
currently it only supports square objects
}
