% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{change_resolution}
\alias{change_resolution}
\title{Title}
\usage{
change_resolution(df_fixations, original_resolution, target_resolution)
}
\arguments{
\item{target_resolution}{}
}
\description{
Title
}
