% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_screen_area_fixations.R
\name{create_areas}
\alias{create_areas}
\title{Returns fixations with added information about to which evnet each fixation belongs}
\usage{
create_areas(height, width)
}
\arguments{
\item{dt_fixations}{fixations data.table. Needs to have posX and posY}

\item{areas}{vector of lists with name and vector of points}
}
\description{
the logic is that two quest events in eyetracker should be separated similarly as two quests in the quest log
if that is correct, we can accept the eyetracking logs and use them to extract fixations for each quest
Even in correct table there are two events that are off, as those are differences between two different sets administered
}
