% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{downsample.eyer}
\alias{downsample.eyer}
\title{Downsamples eyer gaze data}
\usage{
\method{downsample}{eyer}(obj, n, ...)
}
\arguments{
\item{n}{picks every nth recording}

\item{...}{}
}
\value{
EyerObject with downsampled gaze data
}
\description{
Downsamples eyer gaze data
}
