% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{downsample}
\alias{downsample}
\title{Downsamples data}
\usage{
downsample(obj, n, ...)
}
\arguments{
\item{obj}{object to downsample}

\item{n}{picks every nth recording}

\item{...}{}
}
\value{
donwsampled object
}
\description{
Downsamples data
}
