% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-synchronisation.R
\name{eye_synchro_acceptable}
\alias{eye_synchro_acceptable}
\title{This function decides whether the list is acceptable}
\usage{
eye_synchro_acceptable(ls)
}
\arguments{
\item{ls}{list with follwing parameters diff - time duration of the event separation, n_matches - how many good matches were found}
}
\description{
This function decides whether the list is acceptable
}
