% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-synchronisation.R
\name{eye_to_unity_times}
\alias{eye_to_unity_times}
\title{replaces eyetracker_times for quest_times}
\usage{
eye_to_unity_times(fixations, df_sync_times)
}
\arguments{
\item{fixations}{\itemize{
\item this loop might seems weird, as it always modifies all the times regardless of the set
but it actually rewrites only portions of sequencial sets - rewrites 1st with 2nd and 1st and 2nd with 3rd
}}
}
\description{
replaces eyetracker_times for quest_times
}
