% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-object.R
\name{EyerObject}
\alias{EyerObject}
\title{Returns empty eyer object}
\usage{
EyerObject()
}
\value{
eyer object
}
\description{
EyerObject is a list with predetermined fields for consistent functioning of other
functions within the Eyer package. Lots of functions in the eyer package can
work on passed "raw" data, such as fixation dataframes etc., but the benefit of
preprocessing and parsing data inside the eyer object is that the object
is inherently validated and processed and functions know what to expect. This
data science design has been implemented from similar approaches found in Matlab
and Python packages.

EyerObject has the following fields
\itemize{
\item data
\itemize{
\item events
\item fixations
\item diameter
\item gaze
}
\item info: list with eyetracker specific settings, suhc as recording frequency, recorded eye etc.
\itemize{
\item start_time: start time of the first recording in datetime
\item resolution: monitor resolution as a list with width, height fields
\item eyetracker: string with name of the recording device
}
}

Fixations have obligatory columns: time (commonly time since start), x, y, duration
Gaze have obligatory columns: time (commonly time since start), x, y
Events have columns: time, x, y
}
