% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-getters.R
\name{filter_times}
\alias{filter_times}
\title{Filters data between start and end time}
\usage{
filter_times(obj, start, end, ...)
}
\arguments{
\item{obj}{object to filter}

\item{start}{start time}

\item{end}{end time}

\item{...}{}
}
\value{
filtered object
}
\description{
Filters data between start and end time
}
