% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-synchronisation.R
\name{find_better_match}
\alias{find_better_match}
\title{Returns either a list with best synchronised event or NULL if none is found}
\usage{
find_better_match(eye_durations, unity_durations, allowed_difference,
  consecutive)
}
\arguments{
\item{eye_durations}{time separation between each same eye event in milliseconds}

\item{unity_durations}{time separation between unity events in one set}

\item{allowed_difference}{allowed difference between times to be still considered similar}
}
\description{
Returns either a list with best synchronised event or NULL if none is found
}
