% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-synchronisation.R
\name{find_group_match}
\alias{find_group_match}
\title{tries to find a sequency of N elements in eye_durations that correspond to the synchro durations
returns index of first matchin eye event}
\usage{
find_group_match(eye_durations, set_synchro_durations, allowed_difference)
}
\arguments{
\item{allowed_difference}{}
}
\description{
tries to find a sequency of N elements in eye_durations that correspond to the synchro durations
returns index of first matchin eye event
}
