% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-loading.R
\name{find_preprocessed_files}
\alias{find_preprocessed_files}
\title{returns filepaths of preprocessed fixations and events}
\usage{
find_preprocessed_files(folder)
}
\arguments{
\item{folder}{what directory to search in? Expects one fixation and one event to be present}
}
\value{
list with "fixations" and "events" fileds containing filepaths
}
\description{
returns filepaths of preprocessed fixations and events
}
