% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-loading.R
\name{find_unprocessed_file}
\alias{find_unprocessed_file}
\title{Returns list of eyetracker files of given eyetracker}
\usage{
find_unprocessed_file(dir, eyetracker = "SR 1000 ASC")
}
\arguments{
\item{dir}{where to look for}

\item{eyetracker}{what eyetracker file are we dealing with}
}
\description{
Returns list of eyetracker files of given eyetracker
}
