% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-geoms.R
\name{geom_area_boundaries}
\alias{geom_area_boundaries}
\title{Plots area boundaries}
\usage{
geom_area_boundaries(ls, alpha = 0.2, size = 1, color = "black",
  fill = "grey20", ...)
}
\arguments{
\item{ls}{list with xmin, xmax, ymin, ymax fields}

\item{alpha}{rectangle alpha, \strong{default} is 0.2}

\item{size}{rectangle line size, \strong{default} is 1}

\item{color}{rectangle color, \strong{default} is 'black'}

\item{fill}{rectangle fill, \strong{default} is 'grey20'}

\item{...}{optional \code{geom_rect} parameters}
}
\value{
ggplot geom
}
\description{
Plots area boundaries
}
\details{
Wrapper around \code{geom_rect} to be used with \emph{area} functions
}
