% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{geom_eyer_monitor}
\alias{geom_eyer_monitor}
\title{Plots monitor boundaries around data}
\usage{
geom_eyer_monitor(obj, add_centre = FALSE, ...)
}
\arguments{
\item{obj}{EyerObject with \code{obj$info$resolution} field}

\item{add_centre}{if set to TRUE, adds point to the centre of the geom}

\item{...}{ggplot options for geom_rect}
}
\value{
ggplot geom
}
\description{
default geom properties are size = 1, alpha = 0 and color = 'red'. You can overwrite these in teh ...
argument.
}
\details{
creates ggplot geom which plots a geom_rect given specific resolution infromation in the info field
}
