% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-geoms.R
\name{geom_monitor_boundaries}
\alias{geom_monitor_boundaries}
\title{Plots monitor boundaries}
\usage{
geom_monitor_boundaries(width, height, alpha = 0, size = 1.5,
  color = "red", ...)
}
\arguments{
\item{width}{monitor width as a numeric value}

\item{height}{monitor height as a numeric value}

\item{alpha}{rectangle alpha}

\item{size}{rectangle line size}

\item{color}{rectangle color}

\item{...}{optional \code{geom_rect} parameters}
}
\value{
ggplot geom
}
\description{
Plots monitor boundaries
}
\details{
wrapper around \code{geom_area_boundary} with some default values
}
