% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-getters.R
\name{is_data_loaded.eyer}
\alias{is_data_loaded.eyer}
\title{Returns if the eylinker object has both evens and fixations loaded}
\usage{
\method{is_data_loaded}{eyer}(obj)
}
\arguments{
\item{obj}{}
}
\description{
Returns if the eylinker object has both evens and fixations loaded
}
