% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-object.R
\name{is_valid_eyer}
\alias{is_valid_eyer}
\title{Returns if passed object is valid eyer object}
\usage{
is_valid_eyer(obj)
}
\arguments{
\item{obj}{}
}
\description{
Returns if passed object is valid eyer object
}
