% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-loading.R
\name{load_unprocessed_file}
\alias{load_unprocessed_file}
\title{Loads raw file and returns list with evens and fixations}
\usage{
load_unprocessed_file(filepath, eyetracker = "SR 1000")
}
\arguments{
\item{filepath}{path to the file}

\item{eyetracker}{what type of eyetrakcer was used? See readme for instructions}
}
\description{
Loads raw file and returns list with evens and fixations
}
