% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{plot_eye_heatmap}
\alias{plot_eye_heatmap}
\title{Generic function to plot eye heatmap}
\usage{
plot_eye_heatmap(x, y, weights = NULL, ...)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{weights}{optional parameter signifiing durations of x, y coordinates in case they are not of the same importance/ value}

\item{...}{optional ggplot parameters}
}
\value{
ggplot of stat density
}
\description{
Plots heatmap using ggplot stat_density2d function and using x, y and weights coordinates
}
