% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{plot_fixations.eyer}
\alias{plot_fixations.eyer}
\title{Plot fixations and their durations}
\usage{
\method{plot_fixations}{eyer}(obj, duration = T, ...)
}
\arguments{
\item{obj}{Eyer object with fixations loaded}

\item{duration}{if the duration should be plotted. Visualised by color and size. \strong{default} is TRUE}

\item{...}{}
}
\value{
ggplot2 plot
}
\description{
Plot fixations and their durations
}
