% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{plot_fixations}
\alias{plot_fixations}
\title{Plot fixations data}
\usage{
plot_fixations(obj, ...)
}
\arguments{
\item{obj}{object with the fixations data}

\item{...}{}
}
\value{
ggplot2 plot
}
\description{
Plot fixations data
}
