% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{plot_gaze.eyer}
\alias{plot_gaze.eyer}
\title{Plots gaze data}
\usage{
\method{plot_gaze}{eyer}(obj, downsample = 10, ...)
}
\arguments{
\item{obj}{eyer object}

\item{downsample}{picks only every nth element due to gaze data being huge}

\item{...}{}
}
\value{
ggplot2 plot
}
\description{
Plots gaze data
}
