% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{plot_gaze_heatmap.eyer}
\alias{plot_gaze_heatmap.eyer}
\title{Plot heatmap of gaze information}
\usage{
\method{plot_gaze_heatmap}{eyer}(obj, ...)
}
\arguments{
\item{obj}{EyerObject with validgaze data}

\item{...}{additional ggplot arguments}
}
\description{
Plot heatmap of gaze information
}
