% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-visualising.R
\name{plot_gaze_heatmap}
\alias{plot_gaze_heatmap}
\title{Plot heatmap of gaze information}
\usage{
plot_gaze_heatmap(obj, ...)
}
\arguments{
\item{obj}{object with gaze data}

\item{...}{optional ggplot arguments}
}
\value{
ggplot plot
}
\description{
Wrapper around \code{\link{plot_eye_heatmap}} to plot gaze information
}
