% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{preprocess_eye_events}
\alias{preprocess_eye_events}
\title{Reads and cleans events}
\usage{
preprocess_eye_events(df_events, eyetracker)
}
\arguments{
\item{df_events}{data frame with events loaded}

\item{eyetracker}{}
}
\description{
Reads and cleans events
}
