% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-readers.R
\name{read_calibrations}
\alias{read_calibrations}
\title{Reads eyetracker calibration infromation from given file}
\usage{
read_calibrations(filepath, eyetracker)
}
\arguments{
\item{filepath}{path to the raw file}

\item{eyetracker}{which eyetrakcer was used? See readme for instructions}
}
\description{
Reads eyetracker calibration infromation from given file
}
