% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-readers.R
\name{read_eye_events}
\alias{read_eye_events}
\title{Reads fixations from given file}
\usage{
read_eye_events(filepath, eyetracker)
}
\arguments{
\item{eyetracker}{eyetracker type}
}
\description{
Reads fixations from given file
}
