% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-readers.R
\name{read_eye_fixations}
\alias{read_eye_fixations}
\title{Reads fixations from given file}
\usage{
read_eye_fixations(filepath, eyetracker)
}
\arguments{
\item{eyetracker}{eyetracker type}
}
\description{
Reads fixations from given file
}
