% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-readers.R
\name{read_eye_movements}
\alias{read_eye_movements}
\title{Reads movements from given file}
\usage{
read_eye_movements(filepath, eyetracker)
}
\arguments{
\item{filepath}{path to the raw file}

\item{eyetracker}{which eyetrakcer was used? See readme for instructions}
}
\description{
Reads movements from given file
}
