% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-readers.R
\name{read_resolution}
\alias{read_resolution}
\title{Reads eyetracker resolution from given file}
\usage{
read_resolution(filepath, eyetracker)
}
\arguments{
\item{filepath}{path to the raw file}

\item{eyetracker}{which eyetrakcer was used? See readme for instructions}
}
\description{
Reads eyetracker resolution from given file
}
