% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{recalibrate_eye_data}
\alias{recalibrate_eye_data}
\title{Recalibrates eyetracking data}
\usage{
recalibrate_eye_data(obj, new_zero, times = c(), raw_times = FALSE)
}
\arguments{
\item{new_zero}{numeric(2) vector in which to shift the data gets substracted from the x and y axis}

\item{times}{numeric(2) vector to define which part fo the data to select}

\item{raw_times}{commonly, the eyer data times are 0 based witht eh starting time being saved in \code{info$start_time}.
Most operations are then calculated using these 0 based timings. If raw_times is set to TRUE,
data are filtered with ackowledging obj$info$start_time.}
}
\value{
Eyer object with preprocessed data fields
}
\description{
Recalibrates eyetracking data
}
\details{
AS the eyetracking data may get shifted towards certain point,
}
