% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{remove_out_of_bounds.data.frame}
\alias{remove_out_of_bounds.data.frame}
\title{Removes points out of resolution boundary}
\usage{
\method{remove_out_of_bounds}{data.frame}(df, replace, resolution)
}
\arguments{
\item{df}{table with x, y, columns}

\item{replace}{what should the out of bounds values be replaced with? if NULL, rows are delted. NULL by default}

\item{resolution}{defined as a list with width and height in pixesls. e.g: list(width=1920, height=1080)}
}
\value{
eyer object with replaced falues in gaze and fixations
}
\description{
Removes points out of resolution boundary
}
