% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{remove_out_of_bounds.eyer}
\alias{remove_out_of_bounds.eyer}
\title{Removes points out of resolution boundary}
\usage{
\method{remove_out_of_bounds}{eyer}(obj, replace, resolution)
}
\arguments{
\item{obj}{OBject}

\item{replace}{what should the out of bounds values be replaced with? if NULL, rows are delted. NULL by default}

\item{resolution}{efined as a list with width and height in pixesls. e.g: list(width = 1920, height = 1080)}
}
\description{
Removes points out of resolution boundary
}
