% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{remove_out_of_bounds_fixations}
\alias{remove_out_of_bounds_fixations}
\title{Removes fixations out of disp_resolution boundary}
\usage{
remove_out_of_bounds_fixations(df_fixations, disp_resolution = list(width
  = 1920, height = 1080))
}
\arguments{
\item{df_fixations}{fixation table}

\item{disp_resolution}{defined as a list with width and height in pixesls.
Ex: list(width=1920, height=1080)}
}
\description{
Removes fixations out of disp_resolution boundary
}
