% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{remove_outlier_fixations}
\alias{remove_outlier_fixations}
\title{Title}
\usage{
remove_outlier_fixations(df_fixations, disp_resolution = list(width =
  1920, height = 1080))
}
\arguments{
\item{disp_resolution}{}
}
\description{
Title
}
