% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetracker-sr1000.R
\name{SR1000.read_eye_resolution}
\alias{SR1000.read_eye_resolution}
\title{match two digits at least three digit long after Display coords
? signifies non greedy match (as least as possible)}
\usage{
SR1000.read_eye_resolution(path)
}
\description{
match two digits at least three digit long after Display coords
? signifies non greedy match (as least as possible)
}
