% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-synchronisation.R
\name{try_fit_event}
\alias{try_fit_event}
\title{Returns a data frome with synchronising times for each set
Table is in format of set_id, time_eyetracker, time_unity}
\usage{
try_fit_event(eye_event, unity_event, eye_times, unity_times,
  allowed_difference, consecutive = 3)
}
\arguments{
\item{eye_event}{Name of the event in the eyetracker log}

\item{unity_event}{Name of the event in the unity log}

\item{eye_times}{All eyetrackder event times}

\item{unity_times}{All unity event times}

\item{allowed_difference}{How far can two events be separated to be still considered comming from the same source in milliseconds}
}
\description{
Returns a data frome with synchronising times for each set
Table is in format of set_id, time_eyetracker, time_unity
}
