% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{add_screen_area_fixations}
\alias{add_screen_area_fixations}
\title{Adds new column to the fixations data frame with information about a fixation
being within area bounds}
\usage{
add_screen_area_fixations(df_fixations, areas)
}
\arguments{
\item{df_fixations}{fixations}

\item{areas}{list of area lists. Each area list is a list of X and y vectors
of length 2. Ex: list(x=c(0,10), y=c(0,10))}
}
\description{
Adds new column to the fixations data frame with information about a fixation
being within area bounds
}
