% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{change_resolution}
\alias{change_resolution}
\title{Approximates x, y resolution from given coordinates to designated}
\usage{
change_resolution(obj, original, target, ...)
}
\arguments{
\item{obj}{Object to do transformations on}

\item{original}{defined as a list with width and height in pixesls.
Ex: list(width=1920, height=1080)}

\item{target}{defined as a list with width and height in pixesls.
Ex: list(width=1920, height=1080)}
}
\value{
data frame with fixations with changed resolution. Can still yield out of bounds
}
\description{
Approximates x, y resolution from given coordinates to designated
}
