% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-preprocessing.R
\name{change_resolution}
\alias{change_resolution}
\title{Title}
\usage{
change_resolution(df_fixations, original_resolution, target_resolution)
}
\arguments{
\item{df_fixations}{loaded table with fixations}

\item{original_resolution}{defined as a list with width and height in pixesls. Ex: list(width=1920, height=1080)}

\item{target_resolution}{defined as a list with width and height in pixesls. Ex: list(width=1920, height=1080)}
}
\description{
Title
}
