% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-object.R
\name{EyerObject}
\alias{EyerObject}
\title{Returns empty eyer object}
\usage{
EyerObject()
}
\value{
eyer object
}
\description{
EyerObject is a list with predetermined fields for consistent functioning of other
functions within the Eyer package. Lots of functions in the eyer package can
work on passed "raw" data, such as fixation dataframes etc., but the benefit of
preprocessing and parsing data inside the eyer object is that the object
is inherently validated and processed and functions know what to expect. This
data science design has been implemented from similar approaches found in Matlab
and Python packages.

EyerObject has the following fields
\itemize{
\item data
\itemize{
\item events
\item fixations
\item pupil
\item gaze
}
\item start_time: start time of the first recording in datetime
\item resolution: monitor resolution
\item eyetracker: string with name of the recording device
\item settings: list with eyetracker specific settings, suhc as frequency, recorded eye etc.
}

Fixations have an obligatory columns: time (s since start), x, y, duration
}
