% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-getters.R
\name{filter_times.eyer}
\alias{filter_times.eyer}
\title{Filters data between start and end time}
\usage{
\method{filter_times}{eyer}(obj, start, end, data_fields = c(),
  raw_times = F, recalculate_times = F)
}
\arguments{
\item{obj}{Eyer object to filted data from}

\item{start}{start time (inclusive) as is in the time column}

\item{end}{end time (exclusive) as is in the time column}

\item{data_fields}{list of data fields (in obj$data) to filter. If empty, all fields are filtered}

\item{raw_times}{if T, uses obj$info$start_time to calculate times to filter.
Useful if the data have 0 based times since start, but we want to filter based on other values. \strong{default} is FALSE}

\item{recalculate_times}{If true, recalculates times to begin at 0 again. Changes obj$info$start_time. \strong{default} is FALSE}
}
\value{
object with filtered data.
}
\description{
Filters data between start and end time
}
