% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-geoms.R
\name{geom_area_boundaries}
\alias{geom_area_boundaries}
\title{Plots area boundaries}
\usage{
geom_area_boundaries(area, alpha = 0.2, size = 1, color = "black",
  fill = "grey20", ...)
}
\arguments{
\item{alpha}{rectangle alpha}

\item{size}{rectangle line size}

\item{color}{rectangle color}

\item{fill}{rectangle fill}

\item{...}{optional \code{geom_rect} parameters}
}
\value{
ggplot geom
}
\description{
Plots area boundaries
}
\details{
Wrapper around \code{geom_rect} to be used with \emph{area} functions
}
