% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-loading.R
\name{load_eyer_data}
\alias{load_eyer_data}
\title{Loads data from a folder. Loads preprocessed events, gaze and fixations files}
\usage{
load_eyer_data(folder, silent = T, ...)
}
\arguments{
\item{folder}{what directory should be searcher? Not recursive}
}
\value{
EyerObject with filled in data
}
\description{
eyer allows saving of preprocessed data with the function \code{eyer_save()}
this function then allows to load directly those prepared files into its structure
}
