% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-loading.R
\name{open_unprocessed_file}
\alias{open_unprocessed_file}
\title{Finds and loads data from a directory}
\usage{
open_unprocessed_file(dir, eyetracker = "SR 1000 ASC")
}
\arguments{
\item{type}{}
}
\description{
Finds and loads data from a directory
}
