% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-readers.R
\name{read_eye_movements}
\alias{read_eye_movements}
\title{Reads movements from given file}
\usage{
read_eye_movements(filepath, eyetracker)
}
\arguments{
\item{eyetracker}{eyetracker type}
}
\description{
Reads movements from given file
}
