% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyer-saving.R
\name{save_eyer}
\alias{save_eyer}
\title{Saving eyer object into raw or parsed structure}
\usage{
save_eyer(obj, folder = ".", name = "unnamed", robject = FALSE, ...)
}
\arguments{
\item{obj}{eyer object ot save}

\item{folder}{where to save the data. Defaults to working dir. Tries to create the folder if it doesn't exist}

\item{name}{name of the files to be saved}

\item{robject}{if true, then entire object will be saved as an R dataobject.
otherwise fields will be parsed into separate readable csv and json files}

\item{...}{}
}
\description{
Saving eyer object into raw or parsed structure
}
